filesystem_utils/DICOM_Anonymiser/requirements.txt
pydicom
python-magic
pyinstaller